Change Database to DatabaseModel
I implemented several placeholder functions currently. These functions
will allow us to use the Model View Controller design strategy for our
other functins.

The model will be the database.

Soon I will implement some tests that need to be there for the model so
that we can check that all of our functions are working.