Restructure the python folders
This is so that the code can be contained in a better location, where we
can store the tests more correctly.

I'm still not sure if this is the exact way to set it up, but I think it
is a step in the right direction. It lets us now import things from
senseable_gym, instead of each module individually.