Add Machine Objects
This will allow us to abstract information that we get from the
database, as well as have objects that we can interact with, without
having to query the database every time we want to get information.

To be honest, I'm not sure if this is the correct way to do this, but I
think it will work well for us. This way we can use Python objects to do
things, instead of having everything in SQL queries. The results will
only be stored into Machine(s) when necessary.