Start a schema for the sg database
This schema currently doesn't have a whole lot in it, but it is being
incorporated in testing.

Once we are closer to the 'production' product, we will have to be
careful about backing up databases, etc. There are functions that can
wipe the whole database dry, as well as times I'm quite sure I've
destroyed it. That is why I have scripts to run the database from
nothing to something.

I will implement some backup features at a time in the future.