Add support for adding machines to the database
The model can now support adding machines to the database.

I am not sure if we should have th status stored there or something
else. I'm hesitant to put a VERY non-static object in this table, as it
should (at least in my opinion) contain essentially only information
that is true for (nearly) all time for the equipment.