Fix the relationship problems
I'm not 100% sure this is the way it will be implemented in the future
(see the TODO in the database model), but it is working right now and I
think that any speed gains lost from the method I'm currently using
will not every be noticed because they are affecting such small tables
in the databse.

I also left a lot of my previous work commented, so that I could easily
compare it later, or maybe even implement it if I understand the
problems that I was having before. However, at this point, it is working
and so we'll continue on with different areas.