Add ability to save database and multithreading
I am not sure if this is 100% working just yet. I'm hoping to test it on
Dan's branch to see if it's working well or not. Once I've been able to
test it, I will push more commits to fix as necessary. Currently passes
all the tests that I've written it thus far.