Update the reservation class
It now does several error checking measures to make sure that it will be
a valid reservation. There is still more work to be done to check that
no conflicting reservations are being made for the same machine.
However, it now must use machines and users that are in our database.